embed all references to "posixpaths" within _WIN32 tests.
 Unless _WIN32 hold there is no reason to consider "forcing" posix path
specification.  Also, DIALOG_PICKFILE is made to conform to posix yes/no.

+++ src/basic_fun.cpp
@@ -72,8 +72,6 @@ extern "C" char **environ;
 #include "typedefs.hpp"
 #include "base64.hpp"
 #include "objects.hpp"
-//#include "file.hpp"
-
 
 #ifdef HAVE_LOCALE_H
 # include <locale.h>
@@ -2142,7 +2140,7 @@ namespace lib {
       e->Throw( "Search string must be a scalar or one element array: "+
         e->GetParString( 1));
 
-    unsigned long pos = string::npos;
+    unsigned long pos = (unsigned long) string::npos;
     if( nParam > 2)
       {
     BaseGDL* p2 = e->GetParDefined(2);
